setted up typescript configuration for javascript
