--- components/Layout.js
@@ -1,13 +1,0 @@
-import Header from "./Header";
-import Footer from "./Footer";
-import styles from "../styles/Layout.module.scss";
-
-export default function Layout({ children }) {
-  return (
-    <div className={styles.wrapper}>
-      <Header />
-      <main>{children}</main>
-      <Footer />
-    </div>
-  );
-}
