+++ package.json
@@ -13,17 +13,22 @@
   "dependencies": {
     "axios": "^1.6.7",
     "markdown-it": "^14.0.0",
-    "next": "^14.1.0",
+    "next": "^14.2.26",
     "prism-themes": "^1.9.0",
     "prismjs": "^1.29.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "sass": "^1.71.1",
     "sharp": "^0.33.5"
   },
   "devDependencies": {
+    "@types/markdown-it": "^14.1.2",
+    "@types/node": "^22.14.0",
+    "@types/react": "^19.1.0",
+    "@types/react-dom": "^19.1.1",
     "eslint": "^8.57.0",
     "eslint-config-next": "^14.1.0",
-    "prettier": "^3.2.5"
+    "prettier": "^3.2.5",
+    "typescript": "^5.8.3"
   }
 }
