+++ types/blog.ts
@@ -1,0 +1,14 @@
+// types/blog.ts
+
+// types/blog.ts
+export interface Blog {
+  id: string;
+  title: string;
+  description: string;
+  cover: string;
+  created_at: string;
+  tags: string[];
+  // add any other fields your component expects
+}
+
+
